backend/pyproject.toml
[project]
name = "inheir-backend"
version = "0.1.0"
description = "Backend for InHeir.AI"
authors = [
    {name = "InHeir.AI"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {version = "^0.115.6", extras = ["standard"]}
motor = "^3.6.1"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pyjwt = "^2.10.1"
pydantic = {version = "^2.10.5", extras = ["email"]}
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
azure-identity = "^1.19.0"
langchain = "^0.3.20"
openai = {extras = ["azure"], version = "^1.66.3"}
azure-core = "^1.32.0"
azure-ai-generative = "^1.0.0b11"
langchain-openai = "^0.3.8"
langchain-community = "^0.3.19"
azure-ai-evaluation = "1.1.0"
azure-ai-textanalytics = "^5.3.0"
marshmallow = "<3.20"
azure-cognitiveservices-speech = "^1.43.0"
aiofiles = "^24.1.0"
azure-storage-blob = "^12.25.0"
azure-ai-formrecognizer = "^3.3.3"
python-multipart = "^0.0.20"
ffmpeg-python = "^0.2.0"
azure-search-documents = "^11.5.2"
azure-search = "^1.0.0b2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
